refactor: move SessionPluginBundler to shared directory and update imports
